Filters ready for horizontal argument pass
Process now iterates through all the filters in the pipeline. In casse a filter produces FilterTargs (i.e. contain list of packets and list of arg) this list is processed as sub-pipeline by the recursive call of the pipeline processing cycle. The sub-pipeline is processed beginning the first filter after the one which produced the FilterTargs.